requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.3.2
django-el-pagination==3.2.4
django-endless-pagination==2.0
django-filter==2.0.0
django-markdown-deux==1.0.5
django-material==1.2.5
djangorestframework==3.8.2
get==1.0.3
gunicorn==19.9.0
idna==2.7
Markdown==2.6.11
markdown2==2.3.5
post==1.0.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
public==1.0.3
pytz==2018.5
query-string==1.0.2
request==1.0.2
requests==2.19.1
sqlparse==0.2.4
urllib3==1.23
virtualenv==16.0.0
whitenoise==4.0
django-material==1.2.5
